+++ Down/Down/MinusCommand.cpp
@@ -14,8 +14,8 @@ void MinusCommand::execute(VirtualMachine& vm, AbstractFunctionCall& node)
 
 	if (variable1.getTokenType() == Token::TYPE_NUMBER && variable2.getTokenType() == Token::TYPE_NUMBER) {
 
-		int number1 = atoi(variable1.getValue().c_str());
-		int number2 = atoi(variable2.getValue().c_str());
+		double number1 = atof(variable1.getValue().c_str());
+		double number2 = atof(variable2.getValue().c_str());
 
 		vm.setReturnValue(to_string(number1 - number2));
 		vm.setReturnToken(variable1.getTokenType());
